Load Foundation by requiring the SCSS rather than the CSS.
By requiring the CSS, we use Foundation defaults. If we want to customize, we must override it with another CSS. Another shortcoming:  we load styles for all Foundation components while there is a good chance we do not use all components.
By requiring the SCSS, we can customize Foundation to suit our need. We can decide the styles and which components we need.  We also can use SCSS syntax which is more powerful than CSS syntax.
SCSS compilation is already handled by Webpack.